Reworked Lucene indexing to work synchronously as well as asynchronously.
Fixes #5 (issue in cinnamon2-server)

In case of a database rollback, objects and folders will not be indexed (previously this had to be handled in each individual case, which is error-prone). Also the indexing is now triggered whenever the database layer (Hibernate)
updates an Indexable in the database, so the programmer does no longer need to call object.updateIndex() every time an object is changed.